Updated lib and command line application

+++ controller/lib/include/util.h
@@ -67,7 +67,7 @@ namespace avdecc_lib
         /**
          * Convert AEM command status value to its corresponding command status name.
          */
-        AVDECC_CONTROLLER_LIB32_API virtual const char * STDCALL aem_cmd_status_value_to_name(uint16_t aem_cmd_status_value) = 0;
+        AVDECC_CONTROLLER_LIB32_API virtual const char * STDCALL aem_cmd_status_value_to_name(uint32_t aem_cmd_status_value) = 0;
 
         /**
          * Convert command value to its corresponding ACMP command name.
@@ -82,7 +82,7 @@ namespace avdecc_lib
         /**
          * Convert ACMP command status value to its corresponding command status name.
          */
-        AVDECC_CONTROLLER_LIB32_API virtual const char * STDCALL acmp_cmd_status_value_to_name(uint16_t acmp_cmd_status_value) = 0;
+        AVDECC_CONTROLLER_LIB32_API virtual const char * STDCALL acmp_cmd_status_value_to_name(uint32_t acmp_cmd_status_value) = 0;
 
         /**
          * Convert notification value to its corresponding notification name.
